add indicator data

+++ case.go
@@ -4,7 +4,7 @@ import "time"
 
 type Case struct {
 	ID         CaseID    `bson:"_id" json:"id"`
-	Department uint8     `bson:"dep" json:"dep"`
+	Department int       `bson:"dep" json:"dep"`
 	NoticeDate time.Time `bson:"noticeDate" json:"noticeDate"`
 
 	//HospServiceCountReport is the number of hospital services reporting at least one case
@@ -18,5 +18,5 @@ type CaseService interface {
 }
 
 type CaseDAL interface {
-	Get(dep uint8, date time.Time) ([]*Case, error)
+	Get(dep int, date time.Time) ([]*Case, error)
 }
